+++ src/main/java/com/juju/cozyformombackend3/domain/userlog/bloodsugar/dto/request/SaveBloodSugarRecordRequest.java
@@ -1,8 +1,8 @@
 package com.juju.cozyformombackend3.domain.userlog.bloodsugar.dto.request;
 
-import com.juju.cozyformombackend3.domain.userlog.bloodsugar.model.BloodSugarRecordType;
-
 import java.time.LocalDate;
+
+import com.juju.cozyformombackend3.domain.userlog.bloodsugar.model.BloodSugarRecordType;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -14,7 +14,7 @@ public class SaveBloodSugarRecordRequest {
 
 	private LocalDate date;
 	private String type;
-	private double level;
+	private int level;
 
 	public BloodSugarRecordType getType() {
 		return BloodSugarRecordType.valueOf(type);
